towards updated smith examples
